clean up stale invitations

+++ scripts/outside-collaborators-handler.rb
@@ -29,37 +29,6 @@ Signal.trap("TERM") {
 
 
 #########################################################################################
-def get_repo_invitations(repo)
-    loop do
-        check_and_wait_until_reset
-        $client.repository_invitations(repo)
-        rate_limit = $client.rate_limit
-        if rate_limit.remaining > 0 then
-            break
-        end
-    end
-      
-    invitations = []
-
-    last_response = $client.last_response
-    data = last_response.data
-    data.each { |i| invitations << {"id" => i.id,
-                                    "invitee" => i.invitee.login,
-                                    "permissions" => i.permissions} }
-      
-    until last_response.rels[:next].nil?
-        last_response = last_response.rels[:next].get
-        data = last_response.data
-        data.each { |i| invitations << {"id" => i.id,
-                                        "invitee" => i.invitee.login,
-                                        "permissions" => i.permissions} }
-    end
-
-    return invitations
-end
-
-
-#########################################################################################
 def get_repo_collaborators(repo)
     loop do
         check_and_wait_until_reset
@@ -155,27 +124,6 @@ end
 
 
 #########################################################################################
-def repo_member(repo_metadata, groups, user)
-    if repo_metadata then
-        if repo_metadata.key?(user) then
-            return true
-        else
-            repo_metadata.each { |item, props|
-                if (props["type"].casecmp?("group")) then
-                    if groups.key?(item) then
-                        if groups[item].include?(user)
-                            return true
-                        end
-                    end
-                end
-            }
-        end
-    end
-    return false
-end
-
-
-#########################################################################################
 # main
 
 # retrieve information from files
@@ -218,12 +166,11 @@ repos.each { |repo_name, repo_metadata|
         # remove collaborators no longer requested
         get_repo_collaborators(repo_full_name).each { |user|
             check_and_wait_until_reset
-            if !$client.org_member?($org, user) then
-                if !repo_member(repo_metadata, groups, user) then
-                    puts "- Removing collaborator \"#{user}\""
-                    check_and_wait_until_reset
-                    $client.remove_collaborator(repo_full_name, user)
-                end
+            if !$client.org_member?($org, user) &&
+               !repo_member(repo_metadata, groups, user) then
+                puts "- Removing collaborator \"#{user}\""
+                check_and_wait_until_reset
+                $client.remove_collaborator(repo_full_name, user)
             end
         }
 
@@ -231,12 +178,11 @@ repos.each { |repo_name, repo_metadata|
         get_repo_invitations(repo_full_name).each { |invitation|
             invitee = invitation["invitee"]
             check_and_wait_until_reset
-            if !$client.org_member?($org, invitee) then
-                if !repo_member(repo_metadata, groups, invitee) then
-                    puts "- Removing invitee \"#{invitee}\""
-                    check_and_wait_until_reset
-                    $client.delete_repository_invitation(repo_full_name, invitation["id"])
-                end
+            if !$client.org_member?($org, invitee) &&
+               !repo_member(repo_metadata, groups, invitee) then
+                puts "- Removing invitee \"#{invitee}\""
+                check_and_wait_until_reset
+                $client.delete_repository_invitation(repo_full_name, invitation["id"])
             end
         }
